fix: improve global/env translation and cleaning up of overrides (#245)
* Add serviceworker and worker to the list of globalsm to allow flexibility on.

And also improve the tests a bit.

* Add more tests for serviceworker.

* Strip out supersets from environment values in the migrated configs.

This ensures that we aren't providing unnecessary extra values for the `env` key.

I wrote a script to confirm these values were the ones that are supersets, but theoretically we could also just update this to use the `globals` package directly and determine supersets at runtime?

* Add svelte, astro, and vue to match the rust crate.

* Fix the implementation of cleanUpSupersetEnvs.

* Skip envs not in the config.

* Upgrade tsgolint

* Add a many-extends config spec.

This is showing a current problem where multiple extends are not being merged properly, and so we end up with exact-duplicate overrides.

* Add fix for consecutive identical overrides.

Saw this from the mastodon config file after trying to migrate it, so now it's fixed :)

* Add basic cleanup tests.

* Add more complex env superset resolution.

This was heavily helped by AI, but looking through the tests and running it on the snapshots, it looks like it works correctly for all the cases I can come up with.

* Add an integration test that shows the duplicated `"no-console": "off",` overrides being emitted (albeit with differing sets of files).

Will fix this with the next commit.

* Simplify example config/snapshot a lot.

* Fix a comment that gets parsed as an ignore directive

* Add fix for duplicate overrides objects.

If the only difference is the set of files, merge the override objects and combine the file lists.

* Add more test cases.

* update

---------

Co-authored-by: Sysix <sysix@sysix-coding.de>