fix: Fix jsPlugin rules not being disabled properly when set and then disabled in shared configs (#279)
* Add a TODO comment.

* Add test for overriding config merge.

* Update the logic in plugin_rules.ts to fix a bug with the handling of JS plugins and disabled rules.

This one is weird. This *does* drop Mastodon down to "Found 9 warnings and 175 errors." with js plugins, which is good, but I'm not sure this really fixes the underlying problem correctly.

* Add a spec to test this behavior.

* Add another spec.

* Refactor code a bit to use the Category type instead of string for categories.

* Add another snapshot test for overriding config merges that involve a config with a files value.

Currently it does not output the corect final rule settings for this case, and will need to be fixed.

* Fix the logic for the jsPlugins and update the snapshots.

This does, however, result in extra jsPlugins entries for override configs. I may need to fix that by removing unnecessary plugins from these lists.

* Add test for jsplugin override logic.

* Update transformRuleEntry to handle overrides a bit better.

And use options without optional chaining.

Still need to fix the extra jsPlugins being added.

* Fix logic to not add unnecessary extra jsPlugins entries.

May need to add more tests to confirm this works correctly.

* Add tests for isIgnoredPluginRule.

* Add more tests for transformRuleEntry

* Remove deprecated config name from unicorn plugin.

The flat/ prefix is deprecated and this change results in the exact same set of rules.

* Add an "end-to-end" test to validate override merging logic.

I am going to add a few more tests like this to cover various scenarios, rather than relying on the unit tests and snapshots exclusively.

* Add an "e2e" test to check that JS Plugins migrate correctly with overrides.

* Add another typescript snapshot.

This adds a snapshot test for a config file that shows an issue with how we are not merging the typescript plugin into the previous override like we should be, which results in the typescript-eslint rules here not actually being applied.

* Add an e2e test.

Currently it does not pass, but I will update the code to fix things.

* Rewrite the merge logic for overrides to also merge plugins, env, and globals.

Done with a lot of help from GitHub Copilot and the Raptor mini model.

* Revert change to the options param.

* Revert "Rewrite the merge logic for overrides to also merge plugins, env, and globals."

This reverts commit 1ffffd5001c9085de099b4d98d804994e4347b8b.

* Remove override-merging-with-plugins spec for now.